test(token_usage): migrate token usage unit tests to new file and reorganize
